update servis completed
changes-variable-global-handle-login

update mechanic app

+++ Modules/Service/Http/Controllers/PenerimaanServisController.php
@@ -61,6 +61,11 @@ class PenerimaanServisController extends Controller
                                 <i class="fa-solid fa-trash"></i> &nbsp; Delete</a>
                         </li>
                         <li>
+                            <a href="' . url('service/penerimaanServis/create?isEdit=true&id=' . $row->id) . '"
+                                class="dropdown-item d-flex align-items-center btn-edit" data-typemodal="extraLargeModal">
+                                <i class="fa-solid fa-pencil"></i> &nbsp; Edit</a>
+                        </li>
+                        <li>
                             <a href="' . url('service/penerimaanServis/print/' . $row->id . '/penerimaanServis') . '"
                                 class="dropdown-item d-flex align-items-center btn-print">
                                 <i class="fa-solid fa-print"></i> &nbsp; Print Antrian</a>
@@ -85,82 +90,14 @@ class PenerimaanServisController extends Controller
 
     public function show(Request $request, $id)
     {
-        $penerimaanServis = new PenerimaanServis();
-        $row = $penerimaanServis->transaksiServis($id);
-
-        $hargaServis = new HargaServis();
-        $getServis = $hargaServis->getServis()->get();
-
-        foreach ($getServis as $key => $item) {
-            $array_harga_servis[] = [
-                'label' => '<strong>Nama Servis: ' . $item->nama_hargaservis . '</strong> <br />
-                <span>Harga Servis: ' . UtilsHelper::formatUang($item->total_hargaservis) . '</span>',
-                'id' => $item->id,
-            ];
-        }
-        $usersId = Auth::id();
-        $penerimaanServisId = $id;
-
-        $barang = Barang::dataTable()
-            ->with('satuan', 'kategori')
-            ->where('status_barang', 'dijual & untuk servis')
-            ->orWhere('status_barang', 'khusus servis')
-            ->get();
-        $array_barang = [];
-        foreach ($barang as $key => $item) {
-            $array_barang[] = [
-                'id' => $item->id,
-                'label' => '
-            <strong>[' . $item->barcode_barang . '] ' . $item->nama_barang . '</strong> <br />
-            <span>Stok: ' . $item->stok_barang . '</span>
-            '
-            ];
-        }
-        $tipeDiskon = json_encode($this->datastatis['tipe_diskon']);
-        $statusKendaraanServis = ($this->datastatis['status_kendaraan_servis']);
-        $serviceBerkala = ($this->datastatis['servis_berkala']);
-        $cabangId = session()->get('cabang_id');
-
-        $array_status_kendaraan = [];
-        foreach ($statusKendaraanServis as $key => $item) {
-            $array_status_kendaraan[] = [
-                'label' => $item,
-                'id' => $key,
-            ];
-        }
-
-        $array_service_berkala = [];
-        foreach ($serviceBerkala as $key => $item) {
-            $array_service_berkala[] = [
-                'label' => $item,
-                'id' => $key,
-            ];
-        }
-
-        $pesanwa_berkala = $this->datastatis['pesanwa_berkala'];
-        $data = [
-            'row' => $row,
-            'array_harga_servis' => $array_harga_servis,
-            'getServis' => $getServis,
-            'usersId' => $usersId,
-            'penerimaanServisId' => $penerimaanServisId,
-            'barang' => $barang,
-            'array_barang' => $array_barang,
-            'tipeDiskon' => $tipeDiskon,
-            'statusKendaraanServis' => $array_status_kendaraan,
-            'serviceBerkala' => $array_service_berkala,
-            'cabangId' => $cabangId,
-            'pesanwa_berkala' => $pesanwa_berkala,
-        ];
+        $data = UtilsHelper::dataUpdateServis($id, $this->datastatis);
         if ($request->ajax()) {
-            $refresh = $request->input('refresh');
-            if ($refresh) {
-                return response()->json($data);
-            }
-
             $loadData = $request->input('loadData');
             if ($loadData) {
-                return view('service::penerimaanServis.output_detail', $data);
+                return response()->json([
+                    'view' => view('service::penerimaanServis.output_detail', $data)->render(),
+                    'data' => $data,
+                ], 200);
             }
         }
 
@@ -238,11 +175,12 @@ class PenerimaanServisController extends Controller
         $dataUser = json_encode($dataUser);
         $defaultUser = Auth::id();
         $cabangId = session()->get('cabang_id');
-        $isEdit = $request->query('isEdit');
         $penerimaanServis = new PenerimaanServis();
 
         $pesanwa_estimasi = $this->datastatis['pesanwa_estimasi'];
         $totalHutang = 0;
+        $id = $request->input('id');
+        $isEdit = $request->input('isEdit');
         $data = [
             'array_kategori_pembayaran' => $array_kategori_pembayaran,
             'kategoriPembayaran' => $kategoriPembayaran,
@@ -251,7 +189,6 @@ class PenerimaanServisController extends Controller
             'dataUser' => $dataUser,
             'defaultUser' => $defaultUser,
             'cabangId' => $cabangId,
-            'isEdit' => $isEdit,
             'penerimaanServis' => $penerimaanServis,
             'totalHutang' => $totalHutang,
             'array_kategori_servis' => $array_kategori_servis,
@@ -260,6 +197,8 @@ class PenerimaanServisController extends Controller
             'action' => $action,
             'kendaraanServis' => $kendaraanServis,
             'pesanwa_estimasi' => $pesanwa_estimasi,
+            'row' => $penerimaanServis->transaksiServis($id),
+            'isEdit' => $isEdit,
         ];
         if ($request->input('refresh_dataset')) {
             return response()->json($data);
@@ -275,6 +214,13 @@ class PenerimaanServisController extends Controller
      */
     public function store(Request $request)
     {
+        $payloadEdit = $request->input('payloadEdit');
+        $isEdit = $payloadEdit['isEdit'];
+        if ($isEdit) {
+            $penerimaan_servis_id = $payloadEdit['penerimaan_servis_id'];
+            $this->deletePenerimaanServis($penerimaan_servis_id);
+        }
+
         $payloadPenerimaanServis = $request->input('payloadPenerimaanServis');
         $payloadPembayaranServis = $request->input('payloadPembayaranServis');
         $payloadSaldoCustomer = $request->input('payloadSaldoCustomer');
@@ -371,14 +317,13 @@ class PenerimaanServisController extends Controller
      * @param int $id
      * @return Renderable
      */
-    public function destroy($id)
+    private function deletePenerimaanServis($id)
     {
         //
         $penerimaanServis = new PenerimaanServis();
 
         // update if order stock barang
         $getPenerimaanServis = $penerimaanServis->transaksiServis($id);
-
         // update stok barang
         $orderBarang = $getPenerimaanServis->orderBarang;
         if (count($orderBarang) > 0) {
@@ -390,23 +335,43 @@ class PenerimaanServisController extends Controller
             }
         }
 
-
         // saldo customer
-        $saldoDetail = SaldoDetail::where("penerimaan_servis_id", $id)->first();
-        if ($saldoDetail) {
-            $totalDpDetail = $saldoDetail->totalsaldo_detail;
+        $refundSaldo = 0;
+        $pembayaranServis = $getPenerimaanServis->pembayaranServis;
+        foreach ($pembayaranServis as $key => $item) {
+            if (strtolower($item->kategoriPembayaran->nama_kpembayaran) == 'deposit') {
+                $refundSaldo = $item->saldodeposit_pservis - $item->bayar_pservis;
+            }
+        }
 
+        if ($getPenerimaanServis->tanggalambil_pservis != null && $refundSaldo != 0) {
             // update saldo customer
             $customer_id = $getPenerimaanServis->kendaraan->customer_id;
             $saldoCustomer = SaldoCustomer::where('customer_id', $customer_id)->first();
-            $saldoCustomer->jumlah_saldocustomer = $saldoCustomer->jumlah_saldocustomer - $totalDpDetail;
+            $saldoCustomer->jumlah_saldocustomer = $saldoCustomer->jumlah_saldocustomer - $refundSaldo;
             $saldoCustomer->save();
+        } else {
+            $saldoDetail = SaldoDetail::where("penerimaan_servis_id", $id)->first();
+            if ($saldoDetail && $getPenerimaanServis->tanggalambil_pservis == null) {
+                $totalDpDetail = $saldoDetail->totalsaldo_detail;
 
-            // delete saldo detail
-            $saldoDetail->delete();
+                // update saldo customer
+                $customer_id = $getPenerimaanServis->kendaraan->customer_id;
+                $saldoCustomer = SaldoCustomer::where('customer_id', $customer_id)->first();
+                $saldoCustomer->jumlah_saldocustomer = $saldoCustomer->jumlah_saldocustomer - $totalDpDetail;
+                $saldoCustomer->save();
+
+                // delete saldo detail
+                $saldoDetail->delete();
+            }
         }
-
+        // update saldo customer case 2
         PenerimaanServis::destroy($id);
+    }
+
+    public function destroy($id)
+    {
+        $this->deletePenerimaanServis($id);
         return response()->json('Berhasil hapus data', 200);
     }
 
@@ -428,22 +393,31 @@ class PenerimaanServisController extends Controller
 
     public function update(Request $request, $id)
     {
+        $is_pengembalian_servis = $request->input('is_pengembalian_servis');
+        $pengembalian_servis = $request->input('pengembalian_servis');
         $data = $request->except(['_method']);
         if ($data['status_pservis'] == 'proses servis' || $data['status_pservis'] == 'bisa diambil') {
             $tanggal_service_berkala = UtilsHelper::checkTanggalBerkala($data);
             $data = array_merge($data, ['servisberkala_pservis' => $tanggal_service_berkala]);
         }
+        if ($is_pengembalian_servis) {
+            $data = collect($data)
+                ->forget('is_pengembalian_servis')
+                ->forget('pengembalian_servis')
+                ->toArray();
+        }
         PenerimaanServis::find($id)->update($data);
 
-
         $status_pservis = $data['status_pservis'];
-        ServiceHistory::create([
-            'penerimaan_servis_id' => $id,
-            'status_histori' => $status_pservis,
-            'cabang_id' => session()->get('cabang_id'),
-        ]);
+        if (!$is_pengembalian_servis) {
+            ServiceHistory::create([
+                'penerimaan_servis_id' => $id,
+                'status_histori' => $status_pservis,
+                'cabang_id' => session()->get('cabang_id'),
+            ]);
+        }
 
-        // check if cancel
+        // // check if cancel
         $statusServis = ['cancel', 'tidak bisa'];
         if (in_array($status_pservis, $statusServis)) {
 
@@ -477,6 +451,200 @@ class PenerimaanServisController extends Controller
             }
         }
 
-        return response()->json('Berhasil menambahkan progress pengerjaan');
+        $penerimaanServis = new PenerimaanServis();
+        $dataServis = $penerimaanServis->transaksiServis($id);
+
+        // pengembalian servis
+        if ($is_pengembalian_servis) {
+            $penerimaan_servis = $pengembalian_servis['penerimaan_servis'];
+            $nilaigaransi_pservis = $penerimaan_servis['nilaigaransi_pservis'];
+            $tipegaransi_pservis = $penerimaan_servis['tipegaransi_pservis'];
+            $checkTanggalGaransi = UtilsHelper::checkTanggalGaransi($penerimaan_servis);
+            $servisgaransi_pservis = $checkTanggalGaransi;
+
+            // pembayaran servis
+            $pembayaran_servis = $pengembalian_servis['pembayaran_servis'];
+            PembayaranServis::insert($pembayaran_servis);
+
+            // penerimaan servis
+            $penerimaan_servis = $pengembalian_servis['penerimaan_servis'];
+            PenerimaanServis::find($id)->update([
+                'bayar_pservis' => $penerimaan_servis['bayar_pservis'],
+                'hutang_pservis' => $penerimaan_servis['hutang_pservis'],
+                'kembalian_pservis' => $penerimaan_servis['kembalian_pservis'],
+                'nilaigaransi_pservis' => $nilaigaransi_pservis,
+                'tipegaransi_pservis' => $tipegaransi_pservis,
+                'servisgaransi_pservis' => $servisgaransi_pservis,
+                'tanggalambil_pservis' => date('Y-m-d H:i:s'),
+            ]);
+
+            if ($servisgaransi_pservis != null) {
+                $getPenerimaanServis = PenerimaanServis::find($id);
+                $getPenerimaanServis->status_pservis = 'sudah diambil';
+                $getPenerimaanServis->save();
+
+                ServiceHistory::create([
+                    'penerimaan_servis_id' => $id,
+                    'status_histori' => 'sudah diambil',
+                    'cabang_id' => session()->get('cabang_id'),
+                ]);
+            }
+
+            // saldo customer
+            $saldo_customer = $pengembalian_servis['saldo_customer'];
+            $customer_id = $saldo_customer['customer_id'];
+            $saldo_customer_model = SaldoCustomer::where('customer_id', $customer_id)->first();
+            if ($saldo_customer_model) {
+                $saldo_customer_model->update([
+                    'customer_id' => $saldo_customer['customer_id'],
+                    'jumlah_saldocustomer' => $saldo_customer['jumlah_saldocustomer'],
+                    'cabang_id' => $saldo_customer['cabang_id'],
+                ]);
+
+                // saldo detail
+                $saldo_detail = $pengembalian_servis['saldo_detail'];
+                SaldoDetail::create([
+                    'saldo_customer_id' => $saldo_customer_model->id,
+                    'penerimaan_servis_id' => $saldo_detail['penerimaan_servis_id'],
+                    'totalsaldo_detail' => $saldo_detail['totalsaldo_detail'],
+                    'kembaliansaldo_detail' => $saldo_detail['kembaliansaldo_detail'],
+                    'hutangsaldo_detail' => $saldo_detail['hutangsaldo_detail'],
+                    'cabang_id' => $saldo_detail['cabang_id'],
+                ]);
+            }
+        }
+
+        $dataServis = $penerimaanServis->transaksiServis($id);
+        $getPembayaranServis = UtilsHelper::paymentStatisPenerimaanServis($id);
+
+        return response()->json([
+            'message' => 'Berhasil menambahkan progress pengerjaan',
+            'row' => $dataServis,
+            'getPembayaranServis' => $getPembayaranServis,
+        ], 200);
+    }
+
+    // output
+    public function outputUpdateService(Request $request, $id)
+    {
+        $penerimaanServis = new PenerimaanServis();
+        // informasi servis
+        $serviceBerkala = ($this->datastatis['servis_berkala']);
+        $array_service_berkala = [];
+        foreach ($serviceBerkala as $key => $item) {
+            $array_service_berkala[] = [
+                'label' => $item,
+                'id' => $key,
+            ];
+        }
+        $pesanwa_berkala = $this->datastatis['pesanwa_berkala'];
+        $row = $penerimaanServis->transaksiServis($id);
+        $array_kategori_pembayaran = [];
+        $kategoriPembayaran = KategoriPembayaran::dataTable()->where('status_kpembayaran', true)
+            ->get();
+        foreach ($kategoriPembayaran as $value => $item) {
+            $array_kategori_pembayaran[] = [
+                'id' => $item->id,
+                'label' => $item->nama_kpembayaran
+            ];
+        }
+        $array_kategori_pembayaran = json_encode($array_kategori_pembayaran);
+        $getPembayaranServis = json_encode(UtilsHelper::paymentStatisPenerimaanServis($id));
+
+        $loadDataServis = $request->input('loadDataServis');
+        if ($loadDataServis) {
+            $penerimaanServis = new PenerimaanServis();
+            $row = $penerimaanServis->transaksiServis($id);
+            $statusKendaraanServis = ($this->datastatis['status_kendaraan_servis']);
+            $array_status_kendaraan = [];
+            foreach ($statusKendaraanServis as $key => $item) {
+                $array_status_kendaraan[] = [
+                    'label' => $item,
+                    'id' => $key,
+                ];
+            }
+            // servis
+            $data = [
+                'row' => $row,
+                'statusKendaraanServis' => $array_status_kendaraan,
+            ];
+
+            $htmlOrderServis = view('service::penerimaanServis.output.servis', $data)->render();
+
+            // informasi servis
+            $row = $penerimaanServis->transaksiServis($id);
+            $data = [
+                'row' => $row,
+                'statusKendaraanServis' => $array_status_kendaraan,
+                'statusKendaraanServis' => $array_status_kendaraan,
+                'serviceBerkala' => $array_service_berkala,
+                'serviceGaransi' => $array_service_berkala,
+                'pesanwa_berkala' => $pesanwa_berkala,
+                'array_kategori_pembayaran' => $array_kategori_pembayaran,
+            ];
+            $htmlInformasiServis = view('service::penerimaanServis.output.informasiServis', $data)->render();
+
+
+            return response()->json([
+                'row' => $row,
+                'order_servis' => $htmlOrderServis,
+                'informasi_servis' => $htmlInformasiServis,
+                'getPembayaranServis' => $getPembayaranServis
+            ]);
+        }
+
+        $loadDataSparepart = $request->input('loadDataSparepart');
+        if ($loadDataSparepart) {
+            $penerimaanServis = new PenerimaanServis();
+            $row = $penerimaanServis->transaksiServis($id);
+            $statusKendaraanServis = ($this->datastatis['status_kendaraan_servis']);
+            $array_status_kendaraan = [];
+            foreach ($statusKendaraanServis as $key => $item) {
+                $array_status_kendaraan[] = [
+                    'label' => $item,
+                    'id' => $key,
+                ];
+            }
+            // servis
+            $tipeDiskon = ($this->datastatis['tipe_diskon']);
+            $data = [
+                'row' => $row,
+                'tipeDiskon' => $tipeDiskon,
+            ];
+
+            $htmlOrderBarang = view('service::penerimaanServis.output.orderBarang', $data)->render();
+
+            $data = [
+                'row' => $row,
+                'statusKendaraanServis' => $array_status_kendaraan,
+                'serviceBerkala' => $array_service_berkala,
+                'serviceGaransi' => $array_service_berkala,
+                'pesanwa_berkala' => $pesanwa_berkala,
+                'array_kategori_pembayaran' => $array_kategori_pembayaran,
+            ];
+            $htmlInformasiServis = view('service::penerimaanServis.output.informasiServis', $data)->render();
+
+            return response()->json([
+                'row' => $row,
+                'order_barang' => $htmlOrderBarang,
+                'informasi_servis' => $htmlInformasiServis,
+                'getPembayaranServis' => $getPembayaranServis
+            ]);
+        }
+
+        $loadDataHistory = $request->input('loadDataHistory');
+        if ($loadDataHistory) {
+            $penerimaanServis = new PenerimaanServis();
+            // informasi servis
+            $row = $penerimaanServis->transaksiServis($id);
+            $data = [
+                'row' => $row,
+            ];
+            $htmlServiceHistory = view('service::penerimaanServis.output.histori', $data)->render();
+
+            return response()->json([
+                'service_history' => $htmlServiceHistory,
+            ]);
+        }
     }
 }
